file

+++ script.js
@@ -4,10 +4,11 @@ function conversorTemperatura(){
     form.addEventListener('submit', e => e.preventDefault())
 
     var temperatura = document.getElementById('conversor').value
-    tempo = parseFloat((temperatura * 9 / 5) + 32)
+    tempo = parseInt((temperatura * 9 / 5) + 32)
     document.getElementById('info').innerHTML = "Temperatura em Graus Fahrenheit " + tempo + " ºF"
 
 }
 
-
 conversorTemperatura()
+
+
